+++ src/extension.ts
@@ -6,9 +6,9 @@ function activate(context: vscode.ExtensionContext) {
     fixIndentation();
   });
 
-  let netlinx_compile = vscode.commands.registerCommand('extension.netlinx_compile', () => {
-    compileNetlinx();
-  });
+  // let netlinx_compile = vscode.commands.registerCommand('extension.netlinx_compile', () => {
+  //   compileNetlinx();
+  // });
 
   let transfer_command = vscode.commands.registerCommand("extension.netlinx_transfer", () => {
     callShellCommand(vscode.workspace.getConfiguration("netlinx").transferLocation);
@@ -72,7 +72,7 @@ function activate(context: vscode.ExtensionContext) {
   }
 
   context.subscriptions.push(netlinx_format);
-  context.subscriptions.push(netlinx_compile);
+  //context.subscriptions.push(netlinx_compile);      // Removing this to force the use of the build tasks
   context.subscriptions.push(transfer_command);
   context.subscriptions.push(diag_command);
   context.subscriptions.push(help_command);
@@ -101,42 +101,54 @@ function addFolderToWorkspace(folder: string, folderName: string): void {
 }
 
 // This function is the function called by the shortcut on the context menu, or CTRL+F12
-function compileNetlinx(): void {
-  let editor = vscode.window.activeTextEditor;
-  if (!editor) {
-    vscode.window.showErrorMessage("Please open a valid AXS file.");
-    return;
-  }
-  let doc = editor.document;
-  if (doc.languageId === "netlinx-source") {
-    let savedDoc = doc.save();
-    savedDoc.then(() => {
-      let buildCommand = getCompileCommand(doc.fileName);
-      let term = vscode.window.createTerminal('netlinx', vscode.workspace.getConfiguration("netlinx").terminalLocation);
-      term.show();
-      term.sendText(buildCommand);
-    });
-  }
-  else {
-    vscode.window.showErrorMessage("Please open a valid AXS file.");
-    return;
-  }
-}
+// function compileNetlinx(): void {
+//   let editor = vscode.window.activeTextEditor;
+//   if (!editor) {
+//     vscode.window.showErrorMessage("Please open a valid AXS file.");
+//     return;
+//   }
+//   let doc = editor.document;
+//   if (doc.languageId === "netlinx-source") {
+//     let savedDoc = doc.save();
+//     savedDoc.then(() => {
+//       let buildCommand = getCompileCommand(doc.fileName,false);
+//       let term = vscode.window.createTerminal('netlinx', vscode.workspace.getConfiguration("netlinx").terminalLocation);
+//       term.show();
+//       term.sendText(buildCommand);
+//     });
+//   }
+//   else {
+//     vscode.window.showErrorMessage("Please open a valid AXS file.");
+//     return;
+//   }
+// }
 
-function getCompileCommand(fileName: string): string {
+function getCompileCommand(fileName: string, useGlobal: boolean): string {
   let compiler = new NetlinxCompiler();
   compiler.filepaths.push(fileName);
-  if (vscode.workspace.getConfiguration("netlinx").includesLocation.length > 0) {
-    compiler.filepaths.push("-I" + vscode.workspace.getConfiguration("netlinx").includesLocation);
-  }
-  if (vscode.workspace.getConfiguration("netlinx").librariesLocation.length > 0) {
-    compiler.filepaths.push("-L" + vscode.workspace.getConfiguration("netlinx").librariesLocation);
-  }
-  if (vscode.workspace.getConfiguration("netlinx").modulesLocation.length > 0) {
-    compiler.filepaths.push("-M" + vscode.workspace.getConfiguration("netlinx").modulesLocation);
+  if (useGlobal) {
+    if (vscode.workspace.getConfiguration("netlinx").includesLocation.length > 0) {
+      compiler.filepaths.push("-I" + vscode.workspace.getConfiguration("netlinx").includesLocation);
+    }
+    if (vscode.workspace.getConfiguration("netlinx").librariesLocation.length > 0) {
+      compiler.filepaths.push("-L" + vscode.workspace.getConfiguration("netlinx").librariesLocation);
+    }
+    if (vscode.workspace.getConfiguration("netlinx").modulesLocation.length > 0) {
+      compiler.filepaths.push("-M" + vscode.workspace.getConfiguration("netlinx").modulesLocation);
+    }
   }
 
   return compiler.buildCommand();
+}
+
+function countChars(haystack: string, needle: string): number {
+  let count = 0;
+  for (var i = 0; i < haystack.length; i++) {
+    if (haystack[i] === needle){
+      count++;
+    }
+  }
+  return count;
 }
 
 // Code beautifier called by context or keyboard shortcut
@@ -146,6 +158,8 @@ function fixIndentation(): void {
     vscode.window.showErrorMessage("Please open a valid Netlinx file.");
     return;
   }
+
+  // Set up variables for grabbing and replacing the text
   let doc = editor.document;
   var firstLine = doc.lineAt(0);
   var lastLine = doc.lineAt(doc.lineCount - 1);
@@ -153,153 +167,62 @@ function fixIndentation(): void {
     firstLine.range.start.character,
     doc.lineCount - 1,
     lastLine.range.end.character);
+
+  // The indent fixing code
   if (doc.languageId === "netlinx-source" || doc.languageId === "netlinx-include") {
-    let indentLevel = 0;
     let outputText = "";
-    let docText = editor.document.getText();
-    let docLines = docText.split(/\r?\n/);
-    let inComment = 0;    // If we're in a comment and what type
+    let indentLevel = 0;                                        // Current line indent level (number of tabs)
+    let inComment = 0;                                          // If we're in a comment and what level
+    let docText = editor.document.getText();                    // Get the full text of the editor
+    let docLines = docText.split(/\r?\n/);                      // Split into lines
+
+    // Comment weeders
+    let reDeCom1 = /(\/\/.*)/gm;                                // Single line comment
+    let reDeCom2 = /((?:\(\*|\/\*).*(?:\*\)|\*\/))/gm;          // Fully enclosed multiline comment
+    let reDeCom3 = /(.*(?:\*\)|\*\/))/gm;                       // Closing multiline comment
+    let reDeCom4 = /((?:\(\*|\/\*).*)/gm;                       // Opening multiline comment
 
     for (var line = 0; line < docLines.length; line++) {
-      let thisLine = docLines[line].trimLeft();
-      let brOpen = thisLine.indexOf('{');
-      let brClose = thisLine.indexOf('}');
-      let sqOpen = thisLine.indexOf('[');
-      let sqClose = thisLine.indexOf(']');
-      let parOpen = thisLine.indexOf('(');
-      let parClose = thisLine.indexOf(')');
-      let slComment = thisLine.indexOf('//');             // Single line comment
-      let mlComment1Open = thisLine.indexOf('/*');        // Multiline comment 1 open
-      let mlComment1Close = thisLine.indexOf('*/');       // Multiline comment 1 close
-      let mlComment2Open = thisLine.indexOf('(*');        // Multiline comment 2 open
-      let mlComment2Close = thisLine.indexOf('*)');       // Multiline comment 2 close
+      let thisLine = docLines[line];
+      let thisLineTrimmed = docLines[line].trimLeft();
+      let thisLineClean = docLines[line].trimLeft().replace(reDeCom1,"").replace(reDeCom2,"");      // Remove any single line comments and fully enclosed multiline comments
+
+      if (reDeCom3.test(thisLineClean) && inComment > 0) {        // If a multiline comment closes on this line, decrease our comment level
+        inComment = inComment - 1;
+      }
+      if (reDeCom4.test(thisLineClean)) {                         // If a multiline comment opens on this line, increase our comment level
+        inComment = inComment + 1;
+      }
+
+      thisLineClean = thisLineClean.replace(reDeCom3,"").replace(reDeCom4,"");            // Remove any code that we think is inside multiline comments
+      let brOpen = countChars(thisLineClean,'{') - countChars(thisLineClean,'}');         // Check the delta for squiggly brackets
+      let sqOpen = countChars(thisLineClean,'[') - countChars(thisLineClean,']');         // Check the delta for square brackets
+      let parOpen = countChars(thisLineClean,'(') - countChars(thisLineClean,')');        // Check the delta for parenthesis
+      let indentDelta = brOpen + sqOpen + parOpen;                                        // Calculate total delta
 
       // Indent Increase Rules
-      if (inComment > 0) {
-        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        if (mlComment1Close >= 0 || mlComment2Close >= 0) {
-          inComment = 0;
-        }
+      if (inComment > 0) {                                                                // If we're in a multiline comment, just leave the line alone
+        outputText = outputText + thisLine + "\r";
       }
-      else if (slComment === 0) {
-        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
+      else if (indentDelta > 0) {                                                         // If we're increasing indent delta because of this line, the add it, then increase indent
+        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLineTrimmed + "\r";
+        indentLevel = indentLevel + indentDelta;
       }
-      else if (mlComment1Open === 0 || mlComment2Open === 0) {
-        if (mlComment1Open === 0 && mlComment1Close < 0 && mlComment2Close < 0) {
-          inComment = 1;
-        }
-        else if (mlComment2Open === 0 && mlComment1Close < 0 && mlComment2Close < 0) {
-          inComment = 2;
-        }
-        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
+      // If we're decreasing delta, and the line starts with the character that is decreasing it, then decrease first, and then add this line
+      else if (indentDelta < 0 && (thisLineClean[0] === '}' || thisLineClean[0] === ']' || thisLineClean[0] === ')')) {
+        indentLevel = indentLevel + indentDelta;
+        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLineTrimmed + "\r";
       }
-      // Indentation Decrease Rules
-      else if (brClose >= 0 && brOpen < 0) {    // Squiggly Bracket Closing
-        if ((slComment >= 0 && slComment < brClose) ||
-          (mlComment1Open >= 0 && mlComment1Open < brClose && !(mlComment1Close >= 0 && mlComment1Close < brClose)) ||
-          (mlComment2Open >= 0 && mlComment2Open < brClose && !(mlComment2Close >= 0 && mlComment2Close < brClose))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          if (brClose === 0) {    // Check if it's a close bracket on a line by itself; if so, decrease indent on this line
-            console.log('brClose === 0: ' + thisLine)
-            indentLevel = indentLevel - 1;
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          }
-          else {      // There's something else on the line before the close bracket, decrease indent on the next line
-            console.log('brClose != 0: ' + thisLine)
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-            indentLevel = indentLevel - 1;
-          }
-        }
+      else if (indentDelta < 0) {                                                         // If we're decreasing delta but the first character isn't the cause, then we're still inside the block
+        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLineTrimmed + "\r";
+        indentLevel = indentLevel + indentDelta;
       }
-      else if (sqClose >= 0 && sqOpen < 0) {    // Square Bracket Closing
-        if ((slComment >= 0 && slComment < sqClose) ||
-          (mlComment1Open >= 0 && mlComment1Open < sqClose && !(mlComment1Close >= 0 && mlComment1Close < sqClose)) ||
-          (mlComment2Open >= 0 && mlComment2Open < sqClose && !(mlComment2Close >= 0 && mlComment2Close < sqClose))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          if (sqClose === 0) {    // Check if it's a close bracket on a line by itself; if so, decrease indent on this line
-            indentLevel = indentLevel - 1;
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          }
-          else {      // There's something else on the line before the close bracket, decrease indent on the next line
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-            indentLevel = indentLevel - 1;
-          }
-        }
-      }
-      else if (parClose >= 0 && parOpen < 0) {   // Parenthesis closing
-        if ((slComment >= 0 && slComment < parClose) ||
-          (mlComment1Open >= 0 && mlComment1Open < parClose && !(mlComment1Close >= 0 && mlComment1Close < parClose)) ||
-          (mlComment2Open >= 0 && mlComment2Open < parClose && !(mlComment2Close >= 0 && mlComment2Close < parClose))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          if (parClose === 0) {    // Check if it's a close bracket on a line by itself; if so, decrease indent on this line
-            indentLevel = indentLevel - 1;
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          }
-          else {      // There's something else on the line before the close bracket, decrease indent on the next line
-            outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-            indentLevel = indentLevel - 1;
-          }
-        }
-      }
-      else if (brOpen >= 0) {      // Squiggly bracket opening
-        // Make sure there are no comment modifiers that break the opening bracket
-        if ((slComment >= 0 && slComment < brOpen) ||
-          (mlComment1Open >= 0 && mlComment1Open < brOpen && !(mlComment1Close >= 0 && mlComment1Close < brOpen)) ||
-          (mlComment2Open >= 0 && mlComment2Open < brOpen && !(mlComment2Close >= 0 && mlComment2Close < brOpen)) ||
-          (brClose > brOpen && !((slComment >= 0 && slComment < brClose) ||
-            (mlComment1Open >= 0 && mlComment1Open < brClose && !(mlComment1Close >= 0 && mlComment1Close < brClose)) ||
-            (mlComment2Open >= 0 && mlComment2Open < brClose && !(mlComment2Close >= 0 && mlComment2Close < brClose))
-          ))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          indentLevel = indentLevel + 1;
-        }
-      }
-      else if (sqOpen >= 0) {   // Square bracket opening
-        // Make sure there are no comment modifiers that break the opening bracket
-        if ((slComment >= 0 && slComment < sqOpen) ||
-          (mlComment1Open >= 0 && mlComment1Open < sqOpen && !(mlComment1Close >= 0 && mlComment1Close < sqOpen)) ||
-          (mlComment2Open >= 0 && mlComment2Open < sqOpen && !(mlComment2Close >= 0 && mlComment2Close < sqOpen)) ||
-          (sqClose > sqOpen && !((slComment >= 0 && slComment < sqClose) ||
-            (mlComment1Open >= 0 && mlComment1Open < sqClose && !(mlComment1Close >= 0 && mlComment1Close < sqClose)) ||
-            (mlComment2Open >= 0 && mlComment2Open < sqClose && !(mlComment2Close >= 0 && mlComment2Close < sqClose))
-          ))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          indentLevel = indentLevel + 1;
-        }
-      }
-      else if (parOpen >= 0) {    // Parenthesis opening
-        // Make sure there are no comment modifiers that break the opening bracket
-        if ((slComment >= 0 && slComment < parOpen) ||
-          (mlComment1Open >= 0 && mlComment1Open < parOpen && !(mlComment1Close >= 0 && mlComment1Close < parOpen)) ||
-          (mlComment2Open >= 0 && mlComment2Open < parOpen && !(mlComment2Close >= 0 && mlComment2Close < parOpen)) ||
-          (parClose > parOpen && !((slComment >= 0 && slComment < parClose) ||
-            (mlComment1Open >= 0 && mlComment1Open < parClose && !(mlComment1Close >= 0 && mlComment1Close < parClose)) ||
-            (mlComment2Open >= 0 && mlComment2Open < parClose && !(mlComment2Close >= 0 && mlComment2Close < parClose))
-          ))) {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-        }
-        else {
-          outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
-          indentLevel = indentLevel + 1;
-        }
-      }
-
-      else {
-        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLine + "\r";
+      else {                                                                              // indentDelta === 0; do nothing except add the line with the indent
+        outputText = outputText + ('\t'.repeat(indentLevel)) + thisLineTrimmed + "\r";
       }
     };
+
+    // Replace all the code in the editor with the new code
     editor.edit(editBuilder => {
       editBuilder.replace(textRange, outputText);
     });
@@ -326,18 +249,35 @@ async function getCompileTasks(): Promise<vscode.Task[]> {
     let result: vscode.Task[] = [];
     let editor = vscode.window.activeTextEditor;
     let doc = editor.document;
-    let buildCommand = getCompileCommand(doc.fileName);
+    let executable = 'c:\\windows\\system32\\cmd.exe';
+
+    // Create "Local" build, ignoring global folders
+    let buildCommand = getCompileCommand(doc.fileName,false);
 
     let taskDef: NetlinxTaskDefinition = {
       type: 'shell',
-      label: 'Netlinx Build',
+      label: 'Netlinx Build Local',
       buildPath: buildCommand
     }
 
-    let executable = 'c:\\windows\\system32\\cmd.exe';
-
     let command: vscode.ShellExecution = new vscode.ShellExecution(`"${buildCommand}"`, { executable: executable, shellArgs: ['/c'] });
-    let task = new vscode.Task(taskDef, 'Netlinx Build', 'amx-netlinx', command, `$nlrc`);
+    let task = new vscode.Task(taskDef, 'Compile w/Local Files', 'amx-netlinx', command, `$nlrc`);
+    task.definition = taskDef;
+    task.group = vscode.TaskGroup.Build;
+
+    result.push(task);
+
+    // Create "Global" build, including global folders
+    buildCommand = getCompileCommand(doc.fileName,true);
+
+    taskDef = {
+      type: 'shell',
+      label: 'Netlinx Build w/Globals',
+      buildPath: buildCommand
+    }
+
+    command = new vscode.ShellExecution(`"${buildCommand}"`, { executable: executable, shellArgs: ['/c'] });
+    task = new vscode.Task(taskDef, 'Compile w/Global Files', 'amx-netlinx', command, `$nlrc`);
     task.definition = taskDef;
     task.group = vscode.TaskGroup.Build;