+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 0.5.0
+- Added a second build task for compiling with local files only, ignoring the global folders.
+- Removed context shortcut and keyboard shortcut to compile the current file. Please use the Build tasks noted above.
+- Complete rewrite of the beautifier indentation code. Should be much more reliable now.
+
 ### 0.4.3
 - Debugging the new indentation logic